jarfund_backend/requirements.txt
# ─────────────────────────────────────────────────────────────────
#  JarFund Backend — Python Dependencies
# ─────────────────────────────────────────────────────────────────

# ── Core Django ──
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3

# ── Database ──
psycopg2-binary==2.9.10

# ── Authentication ──
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# ── Blockchain / Web3 ──
# web3 7.x ships eth-account 0.13+ as a dependency — no separate pin needed
web3==7.6.1
eth-utils>=5.0.0

# ── Async Task Queue (Celery) ──
celery==5.4.0
redis==5.0.8
django-celery-beat==2.6.0
django-celery-results==2.5.1

# ── API & Serialization ──
djangorestframework-camel-case==1.4.2
drf-spectacular==0.27.2        # OpenAPI schema generation

# ── Caching ──
django-redis==5.4.0

# ── Utilities ──
python-decouple==3.8
Pillow==10.4.0
requests==2.32.3
python-dateutil==2.9.0.post0

# ── Security ──
django-ratelimit==4.1.0

# ── Dev/Testing ──
pytest==8.3.3
pytest-django==4.9.0
pytest-asyncio==0.24.0
factory-boy==3.3.1
faker==28.4.1
coverage==7.6.4
black==24.10.0
isort==5.13.2
flake8==7.1.1

django-extensions==3.2.3
pytest==8.3.3

# ── Production ──
gunicorn==23.0.0
whitenoise==6.7.0
sentry-sdk==2.18.0
